Rename folder & namespace
