+++ app/Helpers/StringHelper.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace App\Helper;
+namespace App\Helpers;
 
 class StringHelper
 {
